requirements.txt
astroid==2.4.2
backcall==0.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
contextvars==2.4
decorator==4.4.2
Flask==1.1.2
google-api-core==1.22.2
google-auth==1.21.1
google-cloud-vision==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
h11==0.9.0
httpcore==0.10.2
httpx==0.14.3
idna==2.10
immutables==0.14
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.5.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pylint==2.6.0
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
rfc3986==1.4.0
rsa==4.6
six==1.15.0
sniffio==1.1.0
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
